Copy stylesheet after build again, reverts dbe55921aec2, fixes #8
Copying the stylesheet before the build fails on a fresh build, because the
build directories aren't created yet.  As this was only done to work around a
bug in the build_sphinx command in Sphinx 1.0.6 and earlier, it is not longer
needed, now that Sphinx 1.0.7 is released.  So lets simply revert the
corresponding changeset dbe55921aec2.